* rcore
- task; obsolete
* timer; periodic task, now is a u64
* state; remove task state

+++ source/main/include/rcore/c_timer.h
@@ -33,7 +33,7 @@ namespace ncore
             task->task_fn     = task_fn;
         }
 
-        void tick_periodic_task(periodic_task_t* task, u32 now_ms);
+        void tick_periodic_task(periodic_task_t* task, u64 now_ms);
 
     }  // namespace ntimer
 }  // namespace ncore
